Fix auto-save issues

+++ inspectors/Saver/SaveManager.js
@@ -23,6 +23,7 @@ class SaveManager
     static T_METHOD = 2;
     static T_FIELD = 3;
     static T_HOOK = 4;
+    static T_HOOK_STATUS = 5;
 
     static __self = null;
 
@@ -37,15 +38,26 @@ class SaveManager
 
     constructor(pContext){
         this.context = pContext;
+        this._ready = false;
         this._enabled = true;
         this._filepath = null;
         this._db = new DB.InMemoryDb(pContext);
 
         this.queue = {
-            method:[],
-            field:[],
-            class:[]
+            method:{},
+            field:{},
+            class:{},
+            hooks: {},
+            "hooks-status": {}
         };
+        this.queueSize = {
+            method:0,
+            field:0,
+            class:0,
+            hooks:0,
+            "hooks-status": 0
+        };
+
 
         // init db
         this.initInternalDB();
@@ -55,6 +67,10 @@ class SaveManager
             this.context.workspace.getSaveDir(),
             SaveManager.FNAME
         );
+    }
+
+    isReady(){
+        return this._ready;
     }
 
     disable(){
@@ -69,11 +85,46 @@ class SaveManager
         return (this._enabled === true);
     }
 
+    addToQueue(pType, pKey, pVal){
+        if(this.queue[pType] !== undefined){
+            this.queue[pType][pKey] = pVal;
+            this.queueSize[pType]++;
+        }
+    }
+
+    removeFromQueue(pType, pKey){
+        let o={};
+
+        if(this.queue[pType] !== undefined){
+            this.queueSize[pType] = 0;
+            for(let i in this.queue[pType]){
+                if(i !== pKey){
+                    o[i] = this.queue[pType][i];
+                    this.queueSize[pType]++;
+                }
+            }
+        }
+        this.queue[pType] = o;
+    }
+
+    getFromQueue(pType, pKey){
+        if(this.queue[pType] !== undefined){
+            if(this.queue[pType][pKey] !== undefined){
+                return this.queue[pType][pKey];
+            }else{
+                return null;
+            }
+        }else{
+            return null;
+        }
+    }
+
     initInternalDB(){
         this._db.newCollection("classes");
         this._db.newCollection("fields");
         this._db.newCollection("methods");
         this._db.newCollection("hooks");
+        this._db.newCollection("hooks-status");
     }
 /*
     newAlias( pType, pObject){
@@ -110,6 +161,9 @@ class SaveManager
     }
 
     
+    updateHookStatus( pHook, pEnable){
+        this._db.getIndex("hooks-status").addEntry(pHook.id, pEnable);
+    }
 
     import(pData){
         let index = null, o=null, hook=null, qflag=0;
@@ -120,8 +174,10 @@ class SaveManager
                 o = this.context.find.get.class(k);
                 if(o != null){
                     o.setAlias(pData.classes.data[k].alias);
+                    this.updateAlias(k, o);
                 }else{
-                    this.queue.class.push(pData.classes.data[k]);
+//                    this.queue.class[k] = pData.classes.data[k];
+                    this.addToQueue("class",k,pData.classes.data[k]);
                     qflag++;
                 }
             }
@@ -133,8 +189,10 @@ class SaveManager
                 o = this.context.find.get.method(k);
                 if(o != null){
                     o.setAlias(pData.methods.data[k].alias);
+                    this.updateAlias(k, o);
                 }else{
-                    this.queue.method.push(pData.methods.data[k]);
+//                    this.queue.method[k] = (pData.methods.data[k]);
+                    this.addToQueue("method",k,pData.methods.data[k]);
                     qflag++;
                 }
             }
@@ -146,8 +204,10 @@ class SaveManager
                 o = this.context.find.get.field(k);
                 if(o != null){
                     o.setAlias(pData.fields.data[k].alias);
+                    this.updateAlias(k, o);
                 }else{
-                    this.queue.field.push(pData.fields.data[k]);
+            //        this.queue.field[k] = (pData.fields.data[k]);
+                    this.addToQueue("field",k,pData.fields.data[k]);
                     qflag++;
                 }
             }
@@ -169,18 +229,46 @@ class SaveManager
 
                     // update the current hook with the imported data
                     hook.updateWith(pData.hooks.data[k].hook, o);
+                    this.updateHook({ hook:hook });
                 }else{
-                    this.queue.hooks.push(pData.hooks.data[k]);
+                   // this.queue.hooks[k] = (pData.hooks.data[k]);
+                    this.addToQueue("hooks",k,pData.hooks.data[k]);
                     qflag++;
                 }
             }
         }
 
-        if(qflag>0)
-            Logger.error("[SAVE] "+qflag+" elements have not imported ...");
-        else
+        if(pData["hooks-status"].size > 0){
+            index = this._db.getIndex("hooks-status");
+            for(let k in pData["hooks-status"].data){
+
+                hook = this.context.hook.getHookByID(k);
+                if(hook != null){
+                    if(pData["hooks-status"].data[k].enable == false){
+                        this.context.bus.prevent("probe.disable");
+                        hook.disable();
+                        this.updateHookStatus(hook, false);
+                    }else{
+                        this.context.bus.prevent("probe.enable");
+                        hook.enable();
+                        this.updateHookStatus(hook, true);
+                    }
+                }else{
+         //           this.queue["hooks-status"][k] = (pData["hooks-status"].data[k]);
+                    this.addToQueue("hooks-status",k,pData.hooks.data[k]);
+                    qflag++;
+                }
+            }
+        }
+
+        if(qflag>0){
+            Logger.error("[SAVE] "+qflag+" elements have not been import. Queued ..");
+        }else
             Logger.info("[SAVE] All elements have imported ...");
         //console.log(this.queue);
+
+        this.context.bus.unprevent("probe.disable")
+        this.context.bus.unprevent("probe.enable")
     }
 
     save(){
@@ -199,6 +287,50 @@ class SaveManager
         data.fields = {data:{}, size:self._db.getIndex("fields").size()};
         data.classes = {data:{}, size:self._db.getIndex("classes").size()};
         data.hooks = {data:{}, size:self._db.getIndex("hooks").size()};
+        data["hooks-status"] = {data:{}, size:self._db.getIndex("hooks-status").size()};
+
+        //if(this.queue.fields)
+
+        if(this.queueSize.fields > 0){
+            for(let i in this.queue.fields){
+                data.fields.data[i] = {
+                    type: SaveManager.T_FIELD,
+                    alias: this.queue.fields[i].alias
+                };
+            }
+        }
+        if(this.queueSize.methods > 0){
+            for(let i in this.queue.methods){
+                data.methods.data[i] = {
+                    type: SaveManager.T_METHOD,
+                    alias: this.queue.methods[i].alias
+                };
+            }
+        }
+        if(this.queueSize.classes > 0){
+            for(let i in this.queue.classes){
+                data.classes.data[i] = {
+                    type: SaveManager.T_CLASS,
+                    alias: this.queue.classes[i].alias
+                };
+            }
+        }
+        if(this.queueSize.hooks > 0){
+            for(let i in this.queue.hooks){
+                data.hooks.data[i] = {
+                    type: SaveManager.T_HOOK,
+                    alias: this.queue.hooks[i].alias
+                };
+            }
+        }
+        if(this.queueSize["hooks-status"] > 0){
+            for(let i in this.queue["hooks-status"]){
+                data["hooks-status"].data[i] = {
+                    type: SaveManager.T_HOOK_STATUS,
+                    enable: this.queue["hooks-status"][i]
+                };
+            }
+        }
 
         if(data.fields.size > 0){
             coll = self._db.getIndex("fields");
@@ -241,6 +373,16 @@ class SaveManager
                 };
             });
         }
+
+        if(data["hooks-status"].size > 0){
+            coll = self._db.getIndex("hooks-status");
+            coll.map(function(ref,obj){
+                data["hooks-status"].data[ref] = {
+                    type: SaveManager.T_HOOK_STATUS,
+                    enable: obj
+                };
+            });
+        }
         
         pData = JSON.stringify(data);
         
@@ -254,30 +396,22 @@ class SaveManager
      */
     write(pFilepath, pData){
         let self = this;
-        _fs_.exists(pFilepath, function(pExists){
-            try{
-                if(_fs_.existsSync(pFilepath+SaveManager.EXT))
-                    _fs_.unlinkSync(pFilepath+SaveManager.EXT);
+        if(_fs_.existsSync(pFilepath)){
+            if(_fs_.existsSync(pFilepath+SaveManager.EXT))
+                 _fs_.unlinkSync(pFilepath+SaveManager.EXT);
 
-                if(pExists)
-                    _fs_.renameSync(pFilepath, pFilepath+SaveManager.EXT);
-
-                _fs_.writeFile(pFilepath, pData, (err)=>{
-                    if(err) 
-                        throw new Error(err);
-
-                    self.context.trigger({
-                        type: "save.write",
-                        data: self
-                    });
-                });
-
-            }catch(e){
-                //console.log(e);
-                Logger.error("[INSPECTOR][SAVE] "+e.msg);
-            }
+            _fs_.renameSync(pFilepath, pFilepath+SaveManager.EXT);
+        }
             
-        });
+        _fs_.writeFile(pFilepath, pData, (err)=>{
+            if(err) 
+                throw new Error(err);
+
+            self.context.trigger({
+                type: "save.write",
+                data: self
+            });
+        });           
     }
 
     restore(){
